refactor: relocate Language primitive from code_audit to engine (#8425)
`Language` (Rust/PHP/JS/TS/Unknown classification + builtin token tables and
per-language predicates) is a source-file primitive with zero dependencies on
higher layers, yet it lived in `code_audit::conventions`. That made `engine`
(edit_op_apply, edit_op_apply::imports) reach up into `code_audit` just to
classify a file — an inverted dependency arrow.

Move the enum + impl (196 lines, no crate:: deps) to `engine/language.rs`, the
foundation layer where audit/refactor/fixer all sit above it. `code_audit::
conventions` re-exports `Language` so the 28 existing consumers keep resolving;
engine consumers are repointed to the canonical `crate::engine::language::Language`.

Severs the engine -> code_audit::Language edge. Behavior-preserving.

Note: a separate engine -> code_audit::AuditFinding edge remains in
refactor_primitive.rs (and edit_op.rs, removed by #8431) — same AuditFinding
misplacement, tracked as follow-up.

Refs #8425
