+++ crates/homeboy-core/src/engine/mod.rs
@@ -18,6 +18,7 @@ pub mod executor;
 pub mod format_write;
 pub mod hooks;
 pub mod invocation;
+pub mod language;
 pub(crate) mod local_files;
 pub mod refactor_primitive;
 pub mod resource;
